+++ data/data-pipeline/tox.ini
@@ -9,9 +9,9 @@ skip_missing_interpreters = true
 # lints python code in src and tests
 basepython = python3.9
 deps = -rrequirements.txt
-commands = black etl application.py config.py utils.py
-           flake8 etl application.py config.py utils.py
-           # pylint etl application.py config.py utils.py
+commands = black data_pipeline
+           flake8 data_pipeline
+           pylint data_pipeline
 
 [testenv:checkdeps]
 # checks the dependencies for security vulnerabilities and open source licenses
